backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.28.15
pydantic==2.1.1
pydantic-settings==2.0.3
alembic==1.11.2
pytest==7.4.0
httpx==0.24.1
python-magic==0.4.27
openai==0.27.8
langchain==0.0.267
tiktoken==0.4.0 